fix(helpers): extract pure functions
Jest has an issue with processing react native, so we moved the pure files out of React Native
